requirements.txt
# Requirements for development and running Riggbot

# Core dependencies
googletrans==4.0.0-rc1
discord.py>=2.2.0
python-dotenv

# Optional
pyinstaller>=5.11
